==> downstream/04_PP_Annot_Cont.Rmd <==
---
title: "04_PP_Annot_Cont"
author: "Ananya Gupta"
date: "2025-11-19"
output: html_document
---

```{r env_loading, results='FALSE',}
# Load packages, data and functions
library(Seurat)
library(tidyverse)
library(SeuratWrappers)
library(magicBatch)
library(BPCells)
library(batchelor)
library(patchwork)
library(Azimuth)
library(plotly)
library(gprofiler2)
library(hdf5r)
library(kableExtra)
library(ggplot2)
library(gridExtra)
library(DT)
library(clustree)
library(Matrix)
library(scales)
library(data.table)
library(dplyr)
library(tidyverse)
library(glue)
library(ggrepel)
library(gprofiler2)
library(multtest)
library(metap)
library(future)
library(parallelly)
library(Cairo)
plan(multicore, workers = 8)


#for parallelization
options(future.globals.maxSize = 8000 * 1024^2) 
#options(future.globals.maxSize = 1e9) was abt 1GB


# Set the random number seed
set.seed(1234)
```


```{r dir setup, include=FALSE}
base_dir <- "PP_outputs"

# Define output directories
rds_dir  <- file.path(base_dir, "PP_all_rep_rds")
fig_dir  <- file.path(base_dir, "PP_all_rep_figs")

# Create folders if they don't exist
dir.create(base_dir, showWarnings = FALSE)
dir.create(rds_dir, showWarnings = FALSE)
dir.create(fig_dir, showWarnings = FALSE)
```
 saving a fig ggsave(
  filename = file.path(fig_dir, "cluster_umap.png"),
  plot = p,
  width = 6,
  height = 5,
  dpi = 300
)


saveRDS(
  object = seu,
  file = file.path(rds_dir, "seu_pp_all_reps.rds")
)


```{r}
## remove the x-axis text and tick
## plot.margin to adjust the white space between each plot.
## ... pass any arguments to VlnPlot in Seurat
modify_vlnplot<- function(obj, 
                          feature, 
                          pt.size = 0, 
                          plot.margin = unit(c(-0.75, 0, -0.75, 0), "cm"),
                          ...) {
  p<- VlnPlot(obj, features = feature, pt.size = pt.size, ... )  + 
    xlab("") + ylab(feature) + ggtitle("") + 
    theme(legend.position = "none", 
          axis.text.x = element_blank(), 
          axis.ticks.x = element_blank(), 
          axis.title.y = element_text(size = rel(1), angle = 0), 
          axis.text.y = element_text(size = rel(1)), 
          plot.margin = plot.margin ) 
  return(p)
}

## extract the max value of the y axis
extract_max<- function(p){
  ymax<- max(ggplot_build(p)$layout$panel_scales_y[[1]]$range$range)
  return(ceiling(ymax))
}


## main function
StackedVlnPlot<- function(obj, features,
                          pt.size = 0, 
                          plot.margin = unit(c(-0.75, 0, -0.75, 0), "cm"),
                          ...) {
  
  plot_list<- purrr::map(features, function(x) modify_vlnplot(obj = obj,feature = x, ...))
  
  # Add back x-axis title to bottom plot. patchwork is going to support this?
  plot_list[[length(plot_list)]]<- plot_list[[length(plot_list)]] +
    theme(axis.text.x=element_text(), axis.ticks.x = element_line())
  
  # change the y-axis tick to only max value 
  ymaxs<- purrr::map_dbl(plot_list, extract_max)
  plot_list<- purrr::map2(plot_list, ymaxs, function(x,y) x + 
                            scale_y_continuous(breaks = c(y)) + 
                            expand_limits(y = y))

  p<- patchwork::wrap_plots(plotlist = plot_list, ncol = 1)
  return(p)
}
```



```{r}
library(Cairo)
save_pub <- function(plot,
                     filename,
                     width = 7,
                     height = 7,
                     dpi = 600) {

  

  # SAVE PDF
  ggsave(
    filename = file.path(fig_dir, paste0(filename, ".pdf")),
    plot = plot,
    device = CairoPDF,
    width = width,
    height = height,
    units = "in"
  )

  # SAVE PNG 
  CairoPNG(
    filename = file.path(fig_dir, paste0(filename, ".png")),
    width  = width * dpi,   # convert inches → pixels
    height = height * dpi,
    res    = dpi
  )

  print(plot) 
  dev.off()

  return(plot)
}

```


```{r}

#Contaminants
lec <- c("Prox1","Lyve1","Pdpn")
lymphocytes <- c("Ptprc","Cd52")
fib_ret <- c("Pdpn","Ccl19","Pdgfra")

# cell markers

hev_markers <- c("St6gal1", "Glycam1", "Chst4", "Ccl21a", "Madcam1")
art_markers <- c("Bmx", "Depp1", "Gja4", "Gja5", "Gkn3", "Sox17")
CRP_markers <- c("Angpt2", "Apln", "Esm1", "Mcam", "Nid2", "Pdgfb", "Pgf", "Vim")
PCV_markers <- c("Ackr1", "Icam1", "Nr2f2", "Sele", "Selp", "Vcam1", "Vwf")
CapEC_marker <- c("Cdh13", "Emcn", "Gja1", "Gpihbp1", "Ly6a", "Ly6c1", "Podxl", "Ramp3")
CapEC2_marker <- c("Atf3", "Cxcl1", "Egr1", "Fos", "Fosb", "Jun", "Junb", "Jund", "Nfkbia", "Nr4a1", "Sgk1") #this overlaps w a lot of others and the capec1 too
CapEC1_marker <- c("Col4a1", "Col4a2", "Hlx", "Id1", "Id3", "Igfbp3")
art_vn_markers <- c("Edn1", "Eln", "Fbln5", "Foxn3", "Klf4", "Ltbp4", "Ptprb")


```


at res 0.2-
0-CapEC2/Capillary (atlas)
1-art
2-Mafb+ Vn
3-PCV
4-HEV
5-CapEC1/(capillary Aqp7+)
6-Pericytes
7-FRC
8-CRP
9-No pecam1, glial cells


```{r}
seu <- readRDS(file.path(rds_dir, "seu_pp_subset_after_umap.rds"))
```

```{r}
seu
```

An object of class Seurat 
23109 features across 14104 samples within 2 assays 
Active assay: RNA (20109 features, 3000 variable features)
 3 layers present: data, counts, scale.data
 1 other assay present: mnn.reconstructed
 3 dimensional reductions calculated: pca, integrated.mnn, umap.mnn
```{r}
seu <- FindClusters(seu, resolution = 0.2, cluster.name = "mnn_clusters")
```


```{r}

saveRDS(
  object = seu,
  file = file.path(rds_dir, "seu_pp_subset_after_res02.rds")) # just in case
  
  
# Load it back
#seu <- readRDS(file.path(rds_dir, "seu_pp_subset_after_res02.rds"))
```

```{r}
  
# Load it back
seu <- readRDS(file.path(rds_dir, "seu_pp_subset_after_res02.rds"))
```


```{r}
head(seu@meta.data)
```

```{r}
a <- DimPlot(seu, group.by="seurat_clusters") + NoLegend() + ggtitle("Cluster Split By Cell") +
  theme(plot.title = element_text(hjust = 0.5))
a <- LabelClusters(a, id = "seurat_clusters",  fontface = "bold", size = 5, bg.colour = "white", bg.r = .2, force = 0)
a
```


```{r, fig.height=6, fig.width=12}

DimPlot(seu, group.by="seurat_clusters", split.by = "condition",label=T) + ggtitle("Viewing Cells in each cluster Split By Condition") +
  theme(plot.title = element_text(hjust = 0.5))#LabelClusters(sp, id = "seurat_clusters",  fontface = "bold", size = 5, bg.colour = "white", bg.r = .2, force = 0)
```
in some cases you will have clusters that do not have enough cells for a particular group - and your function will fail. For these clusters you will need to use FindAllMarkers()


```{r}
annotations <- read.csv("mouse_annotations.csv")
```

```{r}
head(annotations)
```


```{r}
get_conserved <- function(cluster){
  FindConservedMarkers(seu,
                       ident.1 = cluster,
                       grouping.var = "condition",
                       only.pos = TRUE) %>%
    rownames_to_column(var = "gene") %>%
    left_join(y = unique(annotations[, c("gene_name", "description")]),
               by = c("gene" = "gene_name")) %>%
    cbind(cluster_id = cluster, .)
  }
```

```{r}
#map_dfr(inputs_to_function, name_of_function)
levels(Idents(seu))
clusters_to_use <- levels(Idents(seu))
conserved_markers <- map_dfr(clusters_to_use, get_conserved)

```

```{r}

top20_cm <-
  conserved_markers %>%
  mutate(avg_fc = (WT_avg_log2FC + OE_avg_log2FC + KO_avg_log2FC +CRE_CTRL_avg_log2FC) /4) %>% 
  dplyr::group_by(cluster_id) %>%
  dplyr::top_n(n = 20, wt = avg_fc) 
knitr::kable(top20_cm, align = "c")
```

capillary resident regenerative population (CRP) that displays the angiogenic endothelial marker Apln,CRP, for genes associated with endothelial specification during early development or with stem cells (e.g., Ets2, Cxcr4, Nes)


```{r}
write.csv(top20_cm,
          file = "PP_outputs/top20_cm_2.0",
          row.names = FALSE)

```



```{r}
# CRP
FeaturePlot(seu, c("Esm1", "Cdca2", "Apln"),label=T)
```

Venous EC, comprising HEC and non-HEC (Vn) subsets, share expression of the vein-specifying transcription factor Nr2f2 (Coup-TFII)13, and the vein-associated chemokine interceptor Ackr1 (DARC14).

Is 2 Late HEC?

```{r}
# Hec vs hec late
FeaturePlot(seu, c("Glycam1","Ubd", "Plvap"),label=T)
```

```{r}
?FindAllMarkers
```


```{r}
all_markers <- FindAllMarkers(seu, only.pos = TRUE)

```

```{r}
top20 <-
  all_markers %>%
  dplyr::group_by(cluster) %>%
  dplyr::top_n(n = 20, wt = avg_log2FC) 
knitr::kable(top20, align = "c")
```
```{r}
VlnPlot(seu,features="Gpm6b", group.by="seurat_clusters")
```


```{r}
write.csv(top20,
          file = "PP_outputs/top20_allmarkers.csv",
          row.names = FALSE)

```

```{r}
write.csv(all_markers,
          file = "PP_outputs/listof_allmarkers.csv",
          row.names = FALSE)

```



Check identity being stable

```{r}
cluster_2_cells <- subset(seu, subset = seurat_clusters == 2)
Idents(cluster_2_cells) <- "condition"
```

```{r,fig.width=12}
DotPlot(seu, features = c("Nr2f2","Madcam1","Ccl21a","Vcam1","Sele","Icam1","Plvap","St6gal1","Nkx2-3","Mafb"), group.by = "condition")
```

```{r,fig.width=9, fig.height=4}
#Pecam1 = Endothelial Cell
# Sele,Selp=PCV
#Vwf,Icam1,St3gal6= Venous
#Ccl21a,Glycam1,gcnt1 =HEV

DotPlot(seu, features = c("Pecam1","Sele","Selp","Ccl21a","Glycam1","Chst4","Gcnt1","Vwf","Icam1","St3gal6","Lxn"), group.by = "seurat_clusters")
```

```{r,fig.width=8}
DotPlot(seu, features = c("Ubd","Nr2f2","Madcam1","Ccl21a","Vcam1","Sele","Icam1","Podxl","Pecam1"), group.by = "seurat_clusters")
```

Doublet?

```{r,fig.width=10, fig.height=4}
VlnPlot(seu, features = c("nFeature_RNA", "nCount_RNA"), group.by = "seurat_clusters",
              pt.size = 0,
              ncol = 3) #+ ggtitle("After QC")
```

at res 0.2-
0-CapEC2/Capillary (atlas)
1-art
2-Mafb+ Vn
3-PCV
4-HEV
5-CapEC1/(capillary Aqp7+)
6-Pericytes
7-FRC
8-CRP
9-No pecam1, glial cells

```{r}
cell_type  <- c(
  "0" = "CapEC2",
  "1" = "Art",
  "2" = "Mafb+ Vn",
  "3" = "PCV",
  "4" = "HEV",
  "5" = "Aqp7+ CapEC1",
  "6" = "Pericytes",
  "7" = "FRC",
  "8" = "CRP",
  "9" = "Glial Cells")

```

```{r}
seu <- readRDS(file.path(rds_dir, "seu_pp_subset_after_res02.rds"))
```


```{r}
a <- DimPlot(seu, group.by="cluster_annot") + ggtitle("Annotated Peyers Patches") +
  theme(plot.title = element_text(hjust = 0.5))
a <- LabelClusters(a, id = "cluster_annot",  fontface = "bold", size = 2.5, bg.colour = "white", bg.r = .2, force = 0)
a
save_pub(plot=a,filename="annone",width=10)
```


```{r}
# Initialize empty column
seu@meta.data$cluster_annot <- NA

# Map each cluster manually
seu@meta.data$cluster_annot[seu@meta.data$seurat_clusters == 0] <- "CapEC2"
seu@meta.data$cluster_annot[seu@meta.data$seurat_clusters == 1] <- "Artery"
seu@meta.data$cluster_annot[seu@meta.data$seurat_clusters == 2] <- "CapEC Ifn"
seu@meta.data$cluster_annot[seu@meta.data$seurat_clusters == 3] <- "PCV"
seu@meta.data$cluster_annot[seu@meta.data$seurat_clusters == 4] <- "HEV"
seu@meta.data$cluster_annot[seu@meta.data$seurat_clusters == 5] <- "Aqp7+ CapEC1"
seu@meta.data$cluster_annot[seu@meta.data$seurat_clusters == 6] <- "Pericytes"
seu@meta.data$cluster_annot[seu@meta.data$seurat_clusters == 7] <- "FRC"
seu@meta.data$cluster_annot[seu@meta.data$seurat_clusters == 8] <- "CRP"
seu@meta.data$cluster_annot[seu@meta.data$seurat_clusters == 9] <- "Glial Cells"

# Set as identity
Idents(seu) <- "cluster_annot"

# Verify
table(seu@meta.data$cluster_annot)
```


```{r}
#seu <- RenameIdents(seu, new_names)
head(Idents(seu))
```

```{r}


```


```{r}

levels(Idents(seu))
```


```{r}
head(seu@meta.data)
```


```{r}
a <- DimPlot(seu, group.by="cluster_annot") + ggtitle("Annotated Peyers Patches") +
  theme(plot.title = element_text(hjust = 0.5))
a <- LabelClusters(a, id = "cluster_annot",  fontface = "bold", size = 2.5, bg.colour = "white", bg.r = .2, force = 0)
save_pub(plot=a,filename="annone",width=9)
```

```{r}
ncol(seu)
```
```{r}
table(seu$orig.ident)
```

```{r}
# Get cell barcodes to KEEP
cells_to_keep <- WhichCells(seu, expression = cluster_annot %in% c("CapEC2", "Artery", "Mafb+ Vn", "PCV", "HEV", "Aqp7+ CapEC1", "CRP"))

# Subset
seu_filtered <- subset(seu, cells = cells_to_keep)

# Verify
table(seu_filtered$cluster_annot)
```

```{r}
table(seu_filtered$orig.ident)
```
```{r}

saveRDS(
  object = seu_filtered,
  file = file.path(rds_dir, "seu_pp_7BEC.rds")) # just in case

```





```{r}
seu <- seu_filtered
rm(seu_filtered)
```

```{r}
seu
```

Since we subset we need to process again

```{r}
DefaultAssay(seu) <- "RNA"

seu <- NormalizeData(seu)
seu <- FindVariableFeatures(seu, nfeatures = 3000)

#cellcycle
s.genes = gorth(cc.genes.updated.2019$s.genes, source_organism = "hsapiens", target_organism = "mmusculus")$ortholog_name
g2m.genes = gorth(cc.genes.updated.2019$g2m.genes, source_organism = "hsapiens", target_organism = "mmusculus")$ortholog_name

seu <- CellCycleScoring(object=seu, s.features=s.genes, g2m.features=g2m.genes, set.ident=F) #set in the meta.data col phase

seu <- ScaleData(seu,assay="RNA", vars.to.regress = c("S.Score", "G2M.Score"))

seu <- RunPCA(seu, npcs = ifelse(ncol(seu) <= 50, ncol(seu) - 1, 50))
```

```{r}
seu[["RNA"]] <- split(
  seu[["RNA"]],
  f = seu$orig.ident
)
```

```{r}
seu <- IntegrateLayers(seu, method = FastMNNIntegration,
                       features=VariableFeatures(seu),
                       new.reduction = "integrated.mnn",
                       verbose = TRUE)
```

```{r}
seu <- JoinLayers(seu)
```

```{r}
seu <- RunUMAP(seu, reduction = "integrated.mnn", dims = 1:30, reduction.name = "umap.mnn")
seu <- FindNeighbors(seu, reduction = "integrated.mnn", dims = 1:30)
```

```{r}

saveRDS(
  object = seu,
  file = file.path(rds_dir, "seu_pp_7BEC_processed.rds")) # just in case

```


```{r}
seu <- readRDS(file.path(rds_dir, "seu_pp_7BEC_processed.rds"))
```

```{r}
seu <- FindClusters(seu, resolution = 0.4, cluster.name = "mnn_clusters")
```


I need to rename the vn cluster


```{r}
seu <- readRDS(file.path(rds_dir, "seu_pp_7BEC_process_04res.rds"))
```
before

```{r}
ncol(seu)
```



I need to see this mafb thing it might not even be venous but just CAP IFN

```{r}

# cell markers

hev_markers <- c("St6gal1", "Glycam1", "Chst4", "Ccl21a", "Madcam1", "Fut7")
art_markers <- c("Bmx", "Depp1", "Gja4", "Gja5", "Gkn3", "Sox17")
CRP_markers <- c("Angpt2", "Apln", "Esm1", "Mcam", "Nid2", "Pdgfb", "Pgf", "Vim")
PCV_markers <- c("Ackr1", "Icam1", "Nr2f2", "Sele", "Selp", "Vcam1", "Vwf")
CapEC_marker <- c("Cdh13", "Emcn", "Gja1", "Gpihbp1", "Ly6a", "Ly6c1", "Podxl", "Ramp3")
CapEC2_marker <- c("Atf3", "Cxcl1", "Egr1", "Fos", "Fosb", "Jun", "Junb", "Jund", "Nfkbia", "Nr4a1", "Sgk1") #this overlaps w a lot of others and the capec1 too
CapEC1_marker <- c("Col4a1", "Col4a2", "Hlx", "Id1", "Id3", "Igfbp3")
art_vn_markers <- c("Edn1", "Eln", "Fbln5", "Foxn3", "Klf4", "Ltbp4", "Ptprb")

capifn <- c("Irf7","Cxcl9" ,"Cxcl10","Ifit1","Emcn" ,"Ly6a","Cdh13","Igfbp3","Ackr1","Vwf") #Ackr1 nahi hona chaiye in non Vn
```

ankr1 nahi hai,vwf nahi hai
```{r,fig.width=12,fig.height=11}
FeaturePlot(seu,capifn,label=T)
FeaturePlot(seu,CapEC1_marker,label=T)
```



```{r}
table(seu$orig.ident)
```

```{r}
seu@meta.data$cluster_annot[seu@meta.data$cluster_annot == "Barrier EC"] <- "Stressed EC"
```


```{r,fig.width=10, fig.height=4}
x <- VlnPlot(
  object = seu,
  features = "percent.mt",
  group.by = "cluster_annot", pt.size = 0
)
save_pub(plot=x,width = 10, height=4, filename="mtpt_stress_EC" )
```


```{r,fig.width=7,fig.height=5}
an <- DimPlot(seu,group.by="cluster_annot") + ggtitle("BECs of Peyers Patches") +
  theme(plot.title = element_text(hjust = 0.5))
an <- LabelClusters(an, id = "cluster_annot",  fontface = "bold", size = 3, bg.colour = "white", bg.r = .2, force = 0)
an
save_pub(plot=an, filename="becann",width=6,height=5)
#save_pub(plot=an, filename="ppt_clust",width=7,height=5)
```

```{r}
seu@meta.data$cluster_annot[seu@meta.data$cluster_annot == "Vn Sele- Cdh13+"] <- "Vn Sele-"

```

review

```{r,fig.width=14,fig.height=5}
an2 <- DimPlot(seu, split.by= "condition",group.by="cluster_annot") + ggtitle("Annotated Peyers Patches") +
  theme(plot.title = element_text(hjust = 0.5))
LabelClusters(an2, id = "cluster_annot",  fontface = "bold", size = 2.5, bg.colour = "white", bg.r = .2, force = 0)
```
Rename to capecIfn
```{r}
seu@meta.data$cluster_annot[seu@meta.data$cluster_annot == "Mafb+ Vn"] <- "CapEC Ifn"
```

I fear the trec are CapEC1

```{r}
seu@meta.data$cluster_annot[seu@meta.data$cluster_annot == "TrEC"] <- "CapEC1"
```

```{r}

# Verify
table(seu@meta.data$cluster_annot)
```

```{r}
# Initialize empty column
seu@meta.data$cluster_annot <- NA

# Map each cluster manually
seu@meta.data$cluster_annot[seu@meta.data$seurat_clusters == 0] <- "CapEC2"
seu@meta.data$cluster_annot[seu@meta.data$seurat_clusters == 1] <- "Artery"
seu@meta.data$cluster_annot[seu@meta.data$seurat_clusters == 2] <- "CapEC Ifn"
seu@meta.data$cluster_annot[seu@meta.data$seurat_clusters == 3] <- "CapEC1"
seu@meta.data$cluster_annot[seu@meta.data$seurat_clusters == 4] <- "PCV"
seu@meta.data$cluster_annot[seu@meta.data$seurat_clusters == 5] <- "HEV"
seu@meta.data$cluster_annot[seu@meta.data$seurat_clusters == 6] <- "Pre-Artery"
seu@meta.data$cluster_annot[seu@meta.data$seurat_clusters == 7] <- "Vn Sele-"
seu@meta.data$cluster_annot[seu@meta.data$seurat_clusters == 8] <- "Stressed EC"
seu@meta.data$cluster_annot[seu@meta.data$seurat_clusters == 9] <- "CRP"

# Set as identity
Idents(seu) <- "cluster_annot"

# Verify
table(seu@meta.data$cluster_annot)
```


```{r,fig.width=10,fig.height=9}
an2 <- DimPlot(seu, split.by= "condition",group.by="cluster_annot",ncol=2) + ggtitle("Annotated Peyers Patches") +
  theme(plot.title = element_text(hjust = 0.5))
an2 <- LabelClusters(an2, id = "cluster_annot",  fontface = "bold", size = 4, bg.colour = "white", bg.r = .2, force = 0)
an2
save_pub(plot=an2, filename="spcon",width=10,height=9)
```


```{r}

saveRDS(
  object = seu,
  file = file.path(rds_dir, "seu_pp_7BEC_process_04res.rds"))
```

Checking if HEV have higher RNA content

```{r}
seu <- readRDS(file.path(rds_dir, "seu_pp_7BEC_process_04res.rds"))
```

```{r}
seu <- subset(seu, subset = nFeature_RNA > 100 & nCount_RNA < 25000)
```


```{r,fig.width=10, fig.height=4}
p <- VlnPlot(seu, features = c("nCount_RNA"), group.by = "cluster_annot",
              pt.size = 0,
              ncol = 1) #+   geom_boxplot(outlier.shape = NA) +  geom_jitter(width = 0.2)
#+ ggtitle("After QC")
p
# Visualize QC metrics as a violin plot
# VlnPlot(coupko_clean, features = c("nFeature_RNA", "nCount_RNA", "percent.mt"), ncol = 3)
save_pub(plot=p,filename="Rna_of_annots", width=10,height=4)
```
https://divingintogeneticsandgenomics.com/post/stacked-violin-plot-for-visualizing-single-cell-data-in-seurat/
```{r, fig.width==14, fig.height=6}
p <- VlnPlot(
  seu,
  features = c("Ackr1", "Hey1","Nr2f2"),
  split.by = "condition",
  pt.size = 0,idents = c("CapEC2", "Barrier EC","CRP"),
  ncol = 1
)

p & theme(legend.position = "right")  # use patchwork override

```


```{r}
#Diagnostic Resolution
seu_obj <- Seurat::FindClusters(seu, resolution = seq(0.1, 0.6, by=0.1))

```

```{r Find best cluster resolution,fig.height=8,fig.width=10}

clustree::clustree(seu_obj@meta.data[,grep("RNA_snn_res.", colnames(seu_obj@meta.data))],prefix = "RNA_snn_res.", layout="sugiyama")+ggtitle("Finding Best Resolution")+
  theme(
    plot.title = element_text(
      hjust = 0.5,       # centers the title
      margin = margin(b = 12)  # adds space (bottom margin)
    )
  )
```
```{r,fig.height=15,fig.width=7}
Seurat::DimPlot(seu_obj, group.by = c("RNA_snn_res.0.3","RNA_snn_res.0.4","RNA_snn_res.0.5", "RNA_snn_res.0.6"),label=TRUE, ncol=1)
```


```{r}
Idents(seu_obj) <- "RNA_snn_res.0.4"
```

```{r,fig.height=6, fig.width=7}
#art -Bmx, Gja5, Gkn3
#both -Gja4
#pre- Hlx
FeaturePlot(seu_obj,c("Bmx", "Gja4", "Gja5","Gkn3","Hlx","S100a4"),label=T)
```

```{r}
markers_all_bec <- FindAllMarkers(
  seu_obj,
  only.pos = TRUE,
  min.pct = 0.25,
  logfc.threshold = 0.25
)

```

```{r}
write.csv(markers_all_bec,
          file = "PP_outputs/markers_all_bec_res04.csv",
          row.names = FALSE)

```


```{r}
top20_bec <-
  markers_all_bec %>%
  dplyr::group_by(cluster) %>%
  dplyr::top_n(n = 20, wt = avg_log2FC) 
knitr::kable(top20_bec, align = "c")
```


```{r}
write.csv(top20_bec,
          file = "PP_outputs/top20_bec04res.csv",
          row.names = FALSE)

```

```{r}
#hev icam1 is lower in late
FeaturePlot(seu_obj,c("Icam1", "Glycam1","Pf4", "Ccl21a"),label=T)
```


```{r}
annotations <- read.csv("mouse_annotations.csv")
```


```{r}
get_conserved <- function(cluster){
  FindConservedMarkers(seu_obj,
                       ident.1 = cluster,
                       grouping.var = "condition",
                       only.pos = TRUE) %>%
    rownames_to_column(var = "gene") %>%
    left_join(y = unique(annotations[, c("gene_name", "description")]),
               by = c("gene" = "gene_name")) %>%
    cbind(cluster_id = cluster, .)
  }
```



```{r}
#map_dfr(inputs_to_function, name_of_function)
levels(Idents(seu_obj))
clusters_to_use <- levels(Idents(seu_obj))
conserved_markers <- map_dfr(clusters_to_use, get_conserved)
```

```{r}
write.csv(conserved_markers,
          file = "PP_outputs/conserved_markers_bec_res04.csv",
          row.names = FALSE)

```



```{r}
top20_cm_bec <-
  conserved_markers %>%
  mutate(avg_fc = (WT_avg_log2FC + OE_avg_log2FC + KO_avg_log2FC +CRE_CTRL_avg_log2FC) /4) %>% 
  dplyr::group_by(cluster_id) %>%
  dplyr::top_n(n = 20, wt = avg_fc) 
knitr::kable(top20_cm_bec, align = "c")

```

```{r}
write.csv(top20_cm_bec,
          file = "PP_outputs/top20_cm_bec.csv",
          row.names = FALSE)

```


```{r}
FeaturePlot(seu_obj,c("Mafb", "Vwf","Icam1", "St3gal6"),label=T)
```
```{r}
#Sept4 cap

FeaturePlot(seu_obj,c("Id1", "Id3", "Cxcl10"),label=T)
```


```{r,fig.height=6, fig.width=7}
#capec2 -Nr4a1, Sgk1
#Ramp3 ec1 n 2
#
FeaturePlot(seu_obj,c("Ramp3", "Gja1", "Sgk1","Nr4a1"),label=T)
```

```{r}
#crp

FeaturePlot(seu_obj,c("Apoe", "Nuf2", "Apln"),label=T)
```



Five capillary populations shared expression of Cdh13, Emcn, Gja1, and Gpihbp1, previously described as gene markers of capillary EC in PLN6 (Fig. 1f). A transitional phenotype subset (TrEC) expresses canonical capillary genes as well as some HEC genes, albeit at low levels compared to bona fide HEC (Fig. 4a, b). TrEC bridge other CapEC to the venous branch in tSpace projections (Fig. 1e), further suggesting a close relation to HEC. They express Chst2 and the HEC genes St3Gal6 and Fut7, which encode glycosyltransferases for the synthesis of sialyl LewisX (SLex) and 6-sulfo-SLex, carbohydrates that can initiate tethering of lymphocytes under shear flow.Chst4 and Gcnt1 are nearly undetectable in TrECs, suggesting that TrEC and HEC might display different glycotopes. 


```{r,fig.width=9, fig.height=4}
#Pecam1 = Endothelial Cell
# Trec
DotPlot(seu_obj, features = c("Gcnt1","Cdh13","Emcn","Chst4","Gja1","Pecam1"), group.by = "seurat_clusters")
```

```{r}
levels(Idents(seu_obj))
```


```{r,fig.width=9, fig.height=4}
# 7 is Vn but not PCV aka 4
# Trec
Idents(seu_obj) <- "RNA_snn_res.0.4"
DotPlot(seu_obj, features = c("Bgn","Ctsh","Selp","B4galt6","Sele","Ackr1","Cytl1","Gucy1a1","Cdh13","Icam1"), group.by = "RNA_snn_res.0.4")
```


```{r}
#compare daughter clusters

Idents(seu_obj) <- "RNA_snn_res.0.4"

#old clust 0.1 2 become 8 and 5 at res 0.2 

markers_split <- FindMarkers(
  seu,
  ident.1 = "7",
  ident.2 = "4",
  assay = "RNA",      
  logfc.threshold = 0.25,
  min.pct = 0.1
)
head(markers_split %>% arrange(desc(avg_log2FC)), n=20)
```

```{r,fig.width=9, fig.height=4}
#Pecam1 = Endothelial Cell
# Trec
DotPlot(seu_obj, features = c("Pecam1", "Emcn", "Vwf", "Cdh13"), group.by = "RNA_snn_res.0.4")
```
